requirements.txt
# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
supabase>=2.3.0

# AI/ML
openai>=1.12.0
anthropic>=0.18.0
datasets>=2.16.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.26.0

# Development
black>=24.1.0
